Rename `Config` to `ConfigScopes`
