+++ src/lib/gui/ConfigScopes.cpp
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "Config.h"
+#include "ConfigScopes.h"
 
 #include "CommonConfig.h"
 
@@ -68,7 +68,7 @@ void loadWindowsLegacy(QSettings &settings) {
 }
 #endif
 
-Config::Config() {
+ConfigScopes::ConfigScopes() {
   QSettings::setPath(
       QSettings::Format::IniFormat, QSettings::Scope::SystemScope,
       getSystemSettingPath());
@@ -88,20 +88,20 @@ Config::Config() {
   load();
 }
 
-Config::~Config() {
+ConfigScopes::~ConfigScopes() {
   while (!m_pReceievers.empty()) {
     m_pReceievers.pop_back();
   }
 }
 
-void Config::load() {
+void ConfigScopes::load() {
 #if defined(Q_OS_WIN)
   // This call is needed for backwardcapability with old settings.
   loadWindowsLegacy(*m_pSystemSettings);
 #endif
 }
 
-bool Config::hasSetting(const QString &name, Scope scope) const {
+bool ConfigScopes::hasSetting(const QString &name, Scope scope) const {
   switch (scope) {
   case Scope::User:
     return m_pUserSettings->contains(name);
@@ -112,9 +112,9 @@ bool Config::hasSetting(const QString &name, Scope scope) const {
   }
 }
 
-bool Config::isWritable() const { return currentSettings()->isWritable(); }
+bool ConfigScopes::isWritable() const { return currentSettings()->isWritable(); }
 
-QVariant Config::loadSetting(
+QVariant ConfigScopes::loadSetting(
     const QString &name, const QVariant &defaultValue, Scope scope) const {
   switch (scope) {
   case Scope::User:
@@ -126,17 +126,17 @@ QVariant Config::loadSetting(
   }
 }
 
-void Config::setScope(Config::Scope scope) { m_CurrentScope = scope; }
+void ConfigScopes::setScope(ConfigScopes::Scope scope) { m_CurrentScope = scope; }
 
-Config::Scope Config::getScope() const { return m_CurrentScope; }
+ConfigScopes::Scope ConfigScopes::getScope() const { return m_CurrentScope; }
 
-void Config::loadAll() {
+void ConfigScopes::loadAll() {
   for (auto &i : m_pReceievers) {
     i->loadSettings();
   }
 }
 
-void Config::saveAll() {
+void ConfigScopes::saveAll() {
 
   // Save if there are any unsaved changes otherwise skip
   if (unsavedChanges()) {
@@ -151,7 +151,7 @@ void Config::saveAll() {
   }
 }
 
-QSettings *Config::currentSettings() const {
+QSettings *ConfigScopes::currentSettings() const {
   if (m_CurrentScope == Scope::User) {
     return m_pUserSettings.get();
   } else {
@@ -159,11 +159,11 @@ QSettings *Config::currentSettings() const {
   }
 }
 
-void Config::registerReceiever(CommonConfig *receiver) {
+void ConfigScopes::registerReceiever(CommonConfig *receiver) {
   m_pReceievers.push_back(receiver);
 }
 
-bool Config::unsavedChanges() const {
+bool ConfigScopes::unsavedChanges() const {
   if (m_unsavedChanges) {
     return true;
   }
@@ -176,6 +176,6 @@ bool Config::unsavedChanges() const {
   return false;
 }
 
-void Config::markUnsaved() { m_unsavedChanges = true; }
+void ConfigScopes::markUnsaved() { m_unsavedChanges = true; }
 
 } // namespace synergy::gui
